chore: Update ESLint configuration to use .eslintrc.json
This commit updates the ESLint configuration to use the new .eslintrc.json file instead of the deleted .eslintrc.cjs file. The new configuration file includes the necessary plugins and rules for TypeScript and React, as well as the react-refresh plugin. This change improves the maintainability and readability of the ESLint configuration.